+++ src/validate/isStringOk.ts
@@ -1,8 +1,8 @@
 import { makeResponse } from '../schema/utils'
 import { belongsTo } from '../utils/functions'
-import { IStringOptions } from '../utils/types'
+import { StringOptions } from '../utils/types'
 
-export function isStringOk(
+export function isStringOk<T extends string = string>(
   str: any,
   {
     enums,
@@ -10,13 +10,15 @@ export function isStringOk(
     minLength = 1,
     regExp,
     trim = false
-  }: IStringOptions = {}
+  }: StringOptions<T> = {}
 ) {
   if (belongsTo(str, [null, undefined]))
     return makeResponse({ reason: 'Unacceptable value', valid: false })
 
-  if (enums && !belongsTo(str, enums))
-    return makeResponse({ reason: 'Unacceptable value', valid: false })
+  if (enums)
+    return belongsTo(str, enums)
+      ? makeResponse<T>({ valid: true, validated: str })
+      : makeResponse<T>({ reason: 'Unacceptable value', valid: false })
 
   if (regExp && !regExp.test(str))
     return makeResponse({ reason: 'Unacceptable value', valid: false })
@@ -31,5 +33,5 @@ export function isStringOk(
   if (str.length > maxLength)
     return makeResponse({ reason: 'Too long', valid: false })
 
-  return makeResponse<string>({ valid: true, validated: str })
+  return makeResponse<T>({ valid: true, validated: str })
 }